"test" Tiled2Unity exported, reorganized files
Bill's "test" map was exported through Tiled2Unity, and it all looks
good. The map appears to not have any issues, at least with its
appearance.

I also moved the files around a bit, creating an "Other files" folder in
the Unity project folder, meant for anything not directly used in Unity
(for example, Tiled .tmx maps and tilesets are not necessary due to
Tiled2Unity, so they can be placed in that folder).